Spectral Theory of Compact Operators
by The Collective Unconscious

December 31, 2024

Table of contents
1 Introduction   1
2 Spectral Theory of Matrices   1
3 Compact Operators   2
3.1 Statement   2
3.2 Proof   2
Preliminary Lemmas   2
3.3 Invariant Subspaces   3
3.4 Invariant Subspaces   4
3.5 Operators with Compact Power   4
1  Introduction

In functional analysis, compact operators are linear operators on Banach
spaces that map bounded sets to relatively compact sets. In the case of
a Hilbert space H, the compact operators are the closure of the finite
rank operators in the uniform operator topology. In general, operators
on infinite-dimensional spaces feature properties that do not appear in
the finite-dimensional case, i.e. for matrices. The compact operators
are notable in that they share as much similarity with matrices as one
can expect from a general operator. In particular, the spectral
properties of compact operators resemble those of square matrices. This
article first summarizes the corresponding results from the matrix case
before discussing the spectral properties of compact operators. The
reader will see that most statements transfer verbatim from the matrix
case. The spectral theory of compact operators was first developed by F.
Riesz.
2  Spectral Theory of Matrices

The classical result for square matrices is the Jordan canonical form,
which states the following:
Theorem 1. Let A be an n×n complex matrix, i.e. A a linear operator
acting on ℂ^n. If λ_1…*λ_k are the distinct eigenvalues of A, then ℂ^n
can be decomposed into the invariant subspaces of A:
ℂ^n=oplus_(i=1)^kY_i  (1)
The subspace Y_i=Ker(λ_i-A)^m where Ker(λ_i-A)^m=Ker(λ_i-A)^(m+1).
Furthermore, the poles of the resolvent function ζ↦(ζ-A)^(-1) coincide
with the set of eigenvalues of A.
3  Compact Operators

3.1  Statement

Theorem 2. Let X be a Banach space, C be a compact operator acting on X,
and σ(C) be the spectrum of C.
1.Every nonzero λ∈σ(C) is an eigenvalue of C.
2.For all nonzero λ∈σ(C), there exist m such that
Ker((λ-C)^m)=Ker((λ-C)^(m+1)), and this subspace is finite-dimensional.
3.The eigenvalues can only accumulate at 0. If the dimension of X is not
finite, then σ(C) must contain 0.
4.σ(C) is at most countably infinite.
5.Every nonzero λ∈σ(C) is a pole of the resolvent function ζ↦(ζ-C)^(-1).
3.2  Proof

Preliminary Lemmas
The theorem claims several properties of the operator λ-C where λ≠0.
Without loss of generality, it can be assumed that λ=1. Therefore we
consider I-C, I being the identity operator. The proof will require two
lemmas.
Lemma 3. [Riesz's lemma] Let X be a Banach space and Y⊂X, Y≠X, be a
closed subspace. For all ε>0, there exists x∈X such that ‖x‖=1 and:
1-ε≤d(x,Y)≤1  (2)
where d(x,Y) is the distance from x to Y.
This fact will be used repeatedly in the argument leading to the
theorem. Notice that when X is a Hilbert space, the lemma is trivial.
Lemma 4. If C is compact, then Ran(I-C) is closed.
Proof. Let (I-C)*x_n→y in norm. If {x_n} is bounded, then compactness of
C implies that there exists a subsequence x_(n_k) such that C*x_(n_k) is
norm convergent. So x_(n_k)=(I-C)*x_(n_k)+C*x_(n_k) is norm convergent,
to some x. This gives (I-C)*x_(n_k)→(I-C)*x=y.
The same argument goes through if the distances d(x_n,Ker(I-C)) is
bounded. But d(x_n,Ker(I-C)) must be bounded. Suppose this is not the
case. Pass now to the quotient map of (I-C), still denoted by (I-C), on
X/Ker(I-C). The quotient norm on X/Ker(I-C) is still denoted by ‖⋅‖, and
{x_n} are now viewed as representatives of their equivalence classes in
the quotient space.
Take a subsequence {x_(n_k)} such that ‖x_(n_k)‖>k and define a sequence
of unit vectors by z_(n_k)=x_(n_k)/‖x_(n_k)‖. Again we would have
(I-C)*z_(n_k)→(I-C)*z for some z. Since
‖(I-C)*z_(n_k)‖=‖(I-C)*x_(n_k)‖/‖x_(n_k)‖→0, we have (I-C)*z=0 i.e.
z∈Ker(I-C). Since we passed to the quotient map, z=0. This is impossible
because z is the norm limit of a sequence of unit vectors. Thus the
lemma is proven. □
3.3  Invariant Subspaces

As in the matrix case, the above spectral properties lead to a
decomposition of X into invariant subspaces of a compact operator C. Let
λ≠0 be an eigenvalue of C; so λ is an isolated point of σ(C). Using the
holomorphic functional calculus, define the Riesz projection E(λ) by:
E(λ)=(int_γ(ξ-C)^(-1)*d*ξ)/(2*π*i)*  (3)
where γ is a Jordan contour that encloses only λ from σ(C). Let Y be the
subspace Y=E(λ)*X.
C restricted to Y is a compact invertible operator with spectrum {λ},
therefore Y is finite-dimensional. Let ν be such that
Ker(λ-C)^ν=Ker(λ-C)^(ν+1). By inspecting the Jordan form, we see that
(λ-C)^ν=0 while (λ-C)^(ν-1)≠0. The Laurent series of the resolvent
mapping centered at λ shows that:
E(λ)*(λ-C)^ν=(λ-C)^ν*E(λ)=0  (4)
So Y=Ker(λ-C)^ν.
The E(λ) satisfy E(λ)^2=E(λ), so that they are indeed projection
operators or spectral projections. By definition they commute with C.
Moreover E(λ)*E(μ)=0 if λ≠μ.
•  Let X(λ)=E(λ)*X if λ is a non-zero eigenvalue. Thus X(λ) is a
finite-dimensional invariant subspace, the generalised eigenspace of λ.
•  Let X(0) be the intersection of the kernels of the E(λ). Thus X(0) is
a closed subspace invariant under C and the restriction of C to X(0) is
a compact operator with spectrum {0}.
3.4  Invariant Subspaces

As in the matrix case, the above spectral properties lead to a
decomposition of X into invariant subspaces of a compact operator C. Let
λ≠0 be an eigenvalue of C; so λ is an isolated point of σ(C). Using the
holomorphic functional calculus, define the Riesz projection E(λ) by:
E(λ)=(int_γ(ξ-C)^(-1)*d*ξ)/(2*π*i)*  (5)
where γ is a Jordan contour that encloses only λ from σ(C). Let Y be the
subspace Y=E(λ)*X.
C restricted to Y is a compact invertible operator with spectrum {λ},
therefore Y is finite-dimensional. Let ν be such that
Ker(λ-C)^ν=Ker(λ-C)^(ν+1). By inspecting the Jordan form, we see that
(λ-C)^ν=0 while (λ-C)^(ν-1)≠0. The Laurent series of the resolvent
mapping centered at λ shows that:
E(λ)*(λ-C)^ν=(λ-C)^ν*E(λ)=0  (6)
So Y=Ker(λ-C)^ν.
The E(λ) satisfy E(λ)^2=E(λ), so that they are indeed projection
operators or spectral projections. By definition they commute with C.
Moreover E(λ)*E(μ)=0 if λ≠μ.
•  Let X(λ)=E(λ)*X if λ is a non-zero eigenvalue. Thus X(λ) is a
finite-dimensional invariant subspace, the generalised eigenspace of λ.
•  Let X(0) be the intersection of the kernels of the E(λ). Thus X(0) is
a closed subspace invariant under C and the restriction of C to X(0) is
a compact operator with spectrum {0}.
3.5  Operators with Compact Power

If B is an operator on a Banach space X such that B^n is compact for
some n, then the theorem proven above also holds for B.