The Envelope of an Oscillatory Harmonizable Process
by Randall J. Swift

Department of Mathematics
Western Kentucky University
Bowling Green, KY 42101, U.S.A.

Abstract
In 1965, Priestley introduced the class of oscillatory processes and the
concept of their evolutionary spectrum as a tool for the frequency
analysis of these processes. In 1979, Hasofer showed that the envelope
of an oscillatory process does not have a unique representation.
Recently, [Swift1997a] introduced the class of oscillatory harmonizable
processes as an extension of Priestley's oscillatory class. In this
paper, a unique representation for the envelope of a subclass of
oscillatory harmonizable processes is obtained.