On Various Definitions of the Envelope of a Random Process
by R. S. Langley

College of Aeronautics, Cranfield Institute of Technology, Cranfield,
Bedford MK 43 OAL, England

(Received 9 January 1985, and in revised form 7 May 1985)

Abstract
Statistical properties of the envelope definitions of Rice [rice1954],
Crandall and Mark [crandall1963] and Dugundji [dugundji1958] are derived
and compared. It is shown that the definitions of Rice [rice1954] and
Dugundji [dugundji1958] are equivalent, which implies that the envelope
of Rice [rice1954] is independent of the choice of a central frequency.
This contradicts results which have appeared in the literature [lin1967,
lin1976] and the reason for this contradiction is explained. The
envelopes of Crandall and Mark [crandall1963] and Dugundji
[dugundji1958] are found to have the same first order probability
density function but different crossing rates and mean frequencies.