Complex Processes for Envelopes of Normal Noise
by RICHARD ARENS

September 1957

Abstract
The paper presents a brief exposition of the technique of complex normal
random variables as utilized in the study of the envelopes of Gaussian
noise processes. The central concept is the pre-envelope z(t) of a real
normal process. The pre-envelope z(t) of a real function x(t) is a
complex function whose real part is x(t) and whose absolute value is the
envelope, in the sense of high-frequency theory, of x(t). The joint
probability density for z(t), z'(t) is found and used to get the
threshold crossing rate. Consideration of nonstationary processes is
included.