proper transliteration Evolutionary Spectra and Non-Stationary Processes
by M.B. Priestley

February 3rd, 1965

Abstract
We develop an approach to the spectral analysis of non-stationary
processes which is based on the concept of “evolutionary spectra”; that
is, spectral functions which are time dependent, and have a physical
interpretation as local energy distributions over frequency. It is shown
that the notion of evolutionary spectra generalizes the usual definition
of spectra for stationary processes, and that, under certain conditions,
the evolutionary spectrum at each instant of time may be estimated from
a single realization of a process. By such means it is possible to study
processes with continuously changing “spectral patterns”.