Envelopes and Pre-Envelopes of Real Waveforms
by J. DUGUNDJI

University of Southern California, Los Angeles, Calif.
 1958

Abstract
Rice's formula[rice1944] for the “envelope” of a given signal is very
cumbersome; in any case where the signal is not a single sine wave, the
analytical use and explicit calculation of the envelope is practically
prohibitive. A different formula for the envelope is given herein which
is much simpler and easier to handle analytically. We show precisely
that if u^(t) is the Hilbert transform of u(t), then Rice's envelope of
u(t) is the absolute value of the complex-valued function u(t)+i*u^(t).
The function u+i*u^ is called the pre-envelope of u and is shown to be
involved implicitly in some other usual engineering practices.
The Hilbert transform u^ is then studied; it is shown that u^ has the
same power spectrum as u and is uncorrelated with u at the same time
instant. Further, the autocorrelation of the pre-envelope of u is twice
the pre-envelope of the autocorrelation of u.
By using the pre-envelope, the envelope of the output of a linear filter
is easily calculated, and this is used to compute the first probability
density for the envelope of the output of an arbitrary linear filter
when the input is an arbitrary signal plus Gaussian noise. An
application of pre-envelopes to the frequency modulation of an arbitrary
waveform by another arbitrary waveform is also given.
Table of contents
1 Introduction   2
2 Hilbert Transforms   3
3 Definitions and Main Results   4
4 Properties of Pre-Envelopes   4
5 Application to Filters   7
6 Probability Density for Filter Output   9
7 Application to Frequency Modulation   11
Bibliography   13