Injectively Time-Changed Stationary Processes Form a Subclass of Oscillatory Processes
by Stephen Crowley

July 8, 2025

Abstract
Monotonic time changes of stationary Gaussian processes yield
oscillatory processes in the sense of Priestley which are a subclass of
non-stationary Gaussian processes having the spectral representation
K(t,s)=int_(-∞)^∞ϕ_t(λ)ϕ_s(λ)dμ(λ) where μ(λ) is the spectral measure
which has the same interpretation as it does in the case of stationary
processes and the oscillatory function ϕ_t(λ)=e^(i t λ)A_t(λ) where
A_t(λ) is a time-dependent gain function. The monotonicity is necessary
and sufficient for the resulting covariance operator to be self-adjoint
and positive definite. Bidirectional inversion formulas for
reconstructing stationary processes from oscillatory sample paths,
random measure inversion from sample paths, and inversion of the
underslying stationary random white noise measure are derived and the
covariance kernel is expressed as an integral of the ergodic sample
paths of the process.