Stationary and Transient Response Envelopes
Steen Krenk

Henrik O. Madsen

Peter H. Madsen

Abstract
An envelope is introduced by using the Hilbert transform to define a
complex conjugate to the excitation and response processes of a linear
structure. Time-limited stationary excitation is treated in detail, and
the complex correlation function is shown to follow from its stationary
equivalent by use of a suitable differential operator. Simple
expressions are derived for the case of rational spectral density, and a
parametric study of the influence of the frequency content is carried
out. It is found that envelope crossings can be predicted by use of
stationary measures for the mean crossing frequency and the bandwidth of
the response combined with the non-stationary intensity. This is
computationally important as these parameters are often available in
closed form. Finally, the envelope is used to study the first-passage
probability.
Table of contents
1 Introduction   1
2 Stationary Pre-Envelope   3
3 Zero Start Conditions   5
4 Evaluation of Correlation Functions   6
5 Crossing Frequencies   7
6 First Passage Probability   8
7 Summary and Conclusions   9
Bibliography   10