Pre-envelopes of Nonstationary Bandpass Processes
by Harry Urkowitz

Philco Scientific Laboratory, Blue Bell, Pa.

Abstract
Many of the useful properties of pre-envelopes of real waveforms are
shown to hold when the notion of pre-envelope is applied to random
processes which are not wide-sense stationary. In particular, if x(t)
represents a real random process which is not wide-sense stationary and
y(t) is its Hilbert transform, then x(t) and y(t) have the same
autocovariance function and have zero crosscovariance at the same
instant. The autocovariance function of the pre-envelope z(t), given by
z(t)=x(t)+j*y(t), is twice the pre-envelope of the autocovariance
function of x(t).
The notion of a time-dependent power density spectrum allows a simple
interpretation of a bandpass random process which is not wide-sense
stationary. The well-known form of the autocovariance function of a
wide-sense stationary bandpass process carries over with simple changes
to processes which are not wide-sense stationary.
Table of contents
1 Introduction   1
2 The Covariances   2
3 Time Dependent Spectral Densities   3
4 Additional Covariance Properties   5
5 Conclusions   6
References   7
Bibliography   ?