Hilbert Transform Relations for Products
by Gerald D. Cain

Department of Electrical and Electronic Engineering, Polytechnic of
Central London, London W1, England

Abstract
A general formula for the Hilbert transform of a product of
complex-valued functions is developed. Certain simplifications are then
exhibited for products often encountered in the context of modulation
and signal processing. The approach chosen is one of frequency
partitioning; this permits signal definition on complementary sections
of the frequency axis and leads to compact and easily manipulated
expressions.
Table of contents
1 Introduction   1
2 Hilbert Transform Definition   2
3 Bedrosian's Theorem   2
4 Generalization and Analytic Signals   3
5 General Product Formula for Complex-Valued Functions   4
6 Products of Spectrally Disjoint Signals   4
7 Summary   5
Acknowledgment   5
Bibliography   6