Pre-envelopes of Nonstationary Bandpass Processes
by Harry Urkowitz

Philco Scientific Laboratory, Blue Bell, Pa.

Abstract
Many significant properties of pre-envelopes of real waveforms are shown
to extend when the notion of pre-envelope is applied to random processes
that are not wide-sense stationary. Specifically, if x(t) denotes a real
random process which is not wide-sense stationary and y(t) is its
Hilbert transform, then x(t) and y(t) possess the same autocovariance
function and have zero crosscovariance at the same instant. The
autocovariance function of the pre-envelope z(t), given by
z(t)=x(t)+j*y(t), is twice the pre-envelope of the autocovariance
function of x(t).
The introduction of a time-dependent power density spectrum enables a
convenient interpretation of a bandpass random process not assumed to be
wide-sense stationary. The classical form of the autocovariance function
for a wide-sense stationary bandpass process generalizes, with minor
modifications, to nonstationary processes.
Table of contents
1 Introduction   1
2 Covariances   2
3 Time-Dependent Spectral Densities   5
4 Additional Covariance Properties   6
5 Conclusions   8
References   8
Bibliography   8