Riemann's Saddle-point Method and the Riemann-Siegel Formula
by Sir “I've seen hundreds of Graphs Like This Before, Don't Touch it,
Its a Graveyard of Dreams” Berry

Abstract
Riemann's way to calculate his zeta function on the critical line was
based on an application of his saddle-point technique for approximating
integrals that seems astonishing even today. His contour integral for
the remainder in the Dirichlet series for the zeta function involved not
an isolated saddle, nor a saddle near a pole or an end-point or several
coalescing saddles, but the configuration, unfamiliar even now, of a
saddle close to an infinite string of poles. Riemann evaluated the
associated integral exactly, and the resulting Riemann-Siegel formula
underlies ways of computing the Riemann zeros and one of the physical
approaches to the Riemann hypothesis.
2000 Mathematics Subject Classification: 01A55, 11M26, 11Y35, 30B50,
30E15, 34E05, 41A60.
Keywords and Phrases: zeta function, steepest descent, asymptotics.